+++ Braintree/AndroidPayCard.cs
@@ -8,6 +8,7 @@ namespace Braintree
         public string Last4 { get; protected set; }
         public string SourceCardType { get; protected set; }
         public string SourceCardLast4 { get; protected set; }
+        public string SourceDescription { get; protected set; }
         public string VirtualCardType { get; protected set; }
         public string VirtualCardLast4 { get; protected set; }
         public string ExpirationMonth { get; protected set; }
@@ -29,6 +30,7 @@ namespace Braintree
             Last4 = node.GetString("virtual-card-last-4");
             SourceCardLast4 = node.GetString("source-card-last-4");
             VirtualCardLast4 = node.GetString("virtual-card-last-4");
+            SourceDescription = node.GetString("source-description");
             Bin = node.GetString("bin");
             ExpirationMonth = node.GetString("expiration-month");
             ExpirationYear = node.GetString("expiration-year");