+++ Braintree/ApplePayDetails.cs
@@ -6,6 +6,7 @@ namespace Braintree
     {
         public string CardType { get; protected set; }
         public string PaymentInstrumentName { get; protected set; }
+        public string SourceDescription { get; protected set; }
         public string CardholderName { get; protected set; }
         public string ExpirationMonth { get; protected set; }
         public string ExpirationYear { get; protected set; }
@@ -15,6 +16,7 @@ namespace Braintree
         {
             CardType = node.GetString("card-type");
             PaymentInstrumentName = node.GetString("payment-instrument-name");
+            SourceDescription = node.GetString("source-description");
             CardholderName = node.GetString("cardholder-name");
             ExpirationMonth = node.GetString("expiration-month");
             ExpirationYear = node.GetString("expiration-year");