+++ Braintree/Test/Nonce.cs
@@ -12,6 +12,10 @@ namespace Braintree.Test
         public const string ApplePayMastercard = "fake-apple-pay-mastercard-nonce";
         public const string ApplePayAmex = "fake-apple-pay-amex-nonce";
         public const string AndroidPay = "fake-android-pay-nonce";
+        public const string AndroidPayDiscover = "fake-android-pay-discover-nonce";
+        public const string AndroidPayVisa = "fake-android-pay-visa-nonce";
+        public const string AndroidPayMasterCard = "fake-android-pay-mastercard-nonce";
+        public const string AndroidPayAmEx = "fake-android-pay-amex-nonce";
         public const string AbstractTransactable = "fake-abstract-transactable-nonce";
         public const string Europe = "fake-europe-bank-account-nonce";
         public const string Coinbase = "fake-coinbase-nonce";